Initial commit.
To run with enunciate enabled (the build will fail):
mvn clean package

To run without enunciate enabled (for sanity in compilation):
mvn clean package -P'!enunciate'

Effectively what we want to be able to do is to uncomment the
two blocks in the plugin definition and the build will compile.
This requires us to modify the plugin to expose a couple
new configuration parameters and update the plugin version.

To test, you can do a mvn clean install of your updated plugin.
Then, do a mvn clean package of this project.  If this project
builds, then the bug is fixed!